v0.9 updates: 3rd round
New component: lv_hlp
New library: lv_macroable.php
New library: pf_get_debug_type.php
New library: pf_mbstring.php
New library: predis_connect.php
New library: sess_array.php
New test: lv_macroable.php
New test: pf_get_debug_type.php
New test: pf_mbstring.php
New test: predis_connect.php
New test: sess_array.php
Updated component: admin_panel
Updated component: herring
Updated component: login
Updated component: php_polyfill
Updated component: superclosure_router
Updated library: async_process_builder.php
Updated library: lv_arr.php
Updated library: lv_str.php
Updated library: pdo_connect.php
Updated library: pf_json_validate.php
Updated library: sec_captcha.php
Updated library: sec_csrf.php
Updated library: webdevsh.php
Updated README.md
Updated test: async_process_builder.php
Updated test: cache_container.php
Updated test: curl_file_updown.php
Updated test: logger.php
Updated test: lv_arr.php
Updated test: lv_str.php
Updated test: memcached_connect.php
Updated test: ob_cache.php
Updated test: pdo_connect.php
Updated test: pf_json_validate.php
Updated test: queue_worker.php
Updated test: redis_connect.php
Updated test: sec_bruteforce.php
Updated test: sec_captcha.php
Updated test: sec_lv_encrypter.php
Updated test: webdevsh.php
Updated tool: file-watch.php
Updated tool: get-composer.php
Updated tool: strip-php-files.php
Updated tool test: file-watch.php
Updated tool test: serve.php
Updated tool test: websockets.php

+++ bin/tests/serve.php
@@ -135,9 +135,22 @@
 
 		$_serve_test_handler_status=@proc_get_status($_serve_test_handler);
 		if(isset($_serve_test_handler_status['pid']))
+		{
 			@exec('taskkill.exe /F /T /PID '.$_serve_test_handler_status['pid'].' 2>&1');
 
-		proc_terminate($_serve_test_handler);
+			$ch_pid=$_serve_test_handler_status['pid'];
+			$ch_pid_ex=$ch_pid;
+			while(($ch_pid_ex !== null) && ($ch_pid_ex !== ''))
+			{
+				$ch_pid=$ch_pid_ex;
+				$ch_pid_ex=@shell_exec('pgrep -P '.$ch_pid);
+			}
+			if($ch_pid === $_serve_test_handler_status['pid'])
+				proc_terminate($_serve_test_handler);
+			else
+				@exec('kill '.rtrim($ch_pid).' 2>&1');
+		}
+
 		proc_close($_serve_test_handler);
 	}
